dedicated resources: just some reorganization
